requirements.txt
aiohttp==3.12.13
Django==5.2.3
djangorestframework==3.16.0
paho_mqtt==2.1.0
pymodbus==3.9.2
pymongo==3.12.0
python-dotenv==1.1.1
Requests==2.32.4
